Break worker package file into worker package dir
