+++ docs/conf.py
@@ -66,6 +66,8 @@ autodoc_typehints = "description"
 
 autodoc_typehints_description_target = "documented"
 
+autodoc_typehints_format = "short"
+
 autodoc_preserve_defaults = True
 
 autodoc_member_order = "bysource"
