+++ pyproject.toml
@@ -42,6 +42,7 @@ mypy-protobuf = "^3.2.0"
 pydocstyle = "^6.1.1"
 pytest = "^6.2.5"
 pytest-asyncio = "^0.18.0"
+pytest-timeout = "^2.1.0"
 setuptools = "^60.9.3"
 setuptools-rust = "^1.1.2"
 Sphinx = "^4.4.0"
@@ -84,6 +85,8 @@ asyncio_mode = "auto"
 log_cli = true
 log_cli_level = "INFO"
 log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
+timeout = 60
+timeout_func_only = true
 
 [tool.isort]
 profile = "black"
