+++ temporalio/worker/activity.py
@@ -1,11 +1,10 @@
-"""Worker for processing Temporal workflows and/or activities."""
+"""Activity worker."""
 
 from __future__ import annotations
 
 import asyncio
 import concurrent.futures
 import inspect
-import itertools
 import logging
 import multiprocessing
 import multiprocessing.managers
@@ -19,7 +18,6 @@ from datetime import datetime, timedelta, timezone
 from typing import (
     Any,
     Callable,
-    Coroutine,
     Dict,
     Iterable,
     List,
@@ -28,12 +26,10 @@ from typing import (
     Tuple,
     Type,
     Union,
-    cast,
 )
 
 import google.protobuf.duration_pb2
 import google.protobuf.timestamp_pb2
-import typing_extensions
 
 import temporalio.activity
 import temporalio.api.common.v1
@@ -48,101 +44,43 @@ import temporalio.converter
 import temporalio.exceptions
 import temporalio.workflow_service
 
+from .interceptor import (
+    ActivityInboundInterceptor,
+    ActivityOutboundInterceptor,
+    ExecuteActivityInput,
+    Interceptor,
+)
+
 logger = logging.getLogger(__name__)
 
 
-class Worker:
-    """Worker to process workflows and/or activities.
-
-    Once created, workers can be run and shutdown explicitly via :py:meth:`run`
-    and :py:meth:`shutdown`, or they can be used in an ``async with`` clause.
-    """
-
+class _Worker:
     def __init__(
         self,
-        client: temporalio.client.Client,
         *,
+        bridge_worker: Callable[[], temporalio.bridge.worker.Worker],
         task_queue: str,
-        activities: Mapping[str, Callable] = {},
-        activity_executor: Optional[concurrent.futures.Executor] = None,
-        interceptors: Iterable[Interceptor] = [],
-        max_cached_workflows: int = 0,
-        max_concurrent_workflow_tasks: int = 100,
-        max_concurrent_activities: int = 100,
-        max_concurrent_local_activities: int = 100,
-        max_concurrent_wft_polls: int = 5,
-        nonsticky_to_sticky_poll_ratio: float = 0.2,
-        max_concurrent_at_polls: int = 5,
-        no_remote_activities: bool = False,
-        sticky_queue_schedule_to_start_timeout: timedelta = timedelta(seconds=10),
-        max_heartbeat_throttle_interval: timedelta = timedelta(seconds=60),
-        default_heartbeat_throttle_interval: timedelta = timedelta(seconds=30),
-        graceful_shutdown_timeout: timedelta = timedelta(),
-        shared_state_manager: Optional[SharedStateManager] = None,
+        activities: Mapping[str, Callable],
+        activity_executor: Optional[concurrent.futures.Executor],
+        shared_state_manager: Optional[SharedStateManager],
+        type_hint_eval_str: bool,
+        data_converter: temporalio.converter.DataConverter,
+        interceptors: Iterable[Interceptor],
     ) -> None:
-        """Create a worker to process workflows and/or activities.
+        self._bridge_worker = bridge_worker
+        self._task_queue = task_queue
+        self._activity_executor = activity_executor
+        self._shared_state_manager = shared_state_manager
+        self._running_activities: Dict[bytes, _RunningActivity] = {}
+        self._data_converter = data_converter
+        self._interceptors = interceptors
+        # Lazily created on first activity
+        self._worker_shutdown_event: Optional[
+            temporalio.activity._CompositeEvent
+        ] = None
+        self._seen_sync_activity = False
 
-        Args:
-            client: Client to use for this worker. This is required and must be
-                the :py:class:`temporalio.client.Client` instance or have a
-                worker_workflow_service attribute with reference to the original
-                client's underlying service.
-            task_queue: Required task queue for this worker.
-            activities: Mapping of activity type names to activity callables.
-                Activities may be async functions or non-async functions.
-            activity_executor: Concurrent executor to use for non-async
-                activities. This is required if any activities are non-async. If
-                this is a :py:class:`concurrent.futures.ProcessPoolExecutor`,
-                all non-async activities must be picklable.
-            interceptors: Collection of interceptors for this worker. Any
-                interceptors already on the client that also implement
-                :py:class:`Interceptor` are prepended to this list and should
-                not be explicitly given here.
-            max_cached_workflows: If nonzero, workflows will be cached and
-                sticky task queues will be used.
-            max_concurrent_workflow_tasks: Maximum allowed number of workflow
-                tasks that will ever be given to this worker at one time.
-            max_concurrent_activities: Maximum number of activity tasks that
-                will ever be given to this worker concurrently.
-            max_concurrent_local_activities: Maximum number of local activity
-                tasks that will ever be given to this worker concurrently.
-            max_concurrent_wft_polls: Maximum number of concurrent poll workflow
-                task requests we will perform at a time on this worker's task
-                queue.
-            nonsticky_to_sticky_poll_ratio: max_concurrent_wft_polls * this
-                number = the number of max pollers that will be allowed for the
-                nonsticky queue when sticky tasks are enabled. If both defaults
-                are used, the sticky queue will allow 4 max pollers while the
-                nonsticky queue will allow one. The minimum for either poller is
-                1, so if ``max_concurrent_wft_polls`` is 1 and sticky queues are
-                enabled, there will be 2 concurrent polls.
-            max_concurrent_at_polls: Maximum number of concurrent poll activity
-                task requests we will perform at a time on this worker's task
-                queue.
-            no_remote_activities: If true, this worker will only handle workflow
-                tasks and local activities, it will not poll for activity tasks.
-            sticky_queue_schedule_to_start_timeout: How long a workflow task is
-                allowed to sit on the sticky queue before it is timed out and
-                moved to the non-sticky queue where it may be picked up by any
-                worker.
-            max_heartbeat_throttle_interval: Longest interval for throttling
-                activity heartbeats.
-            default_heartbeat_throttle_interval: Default interval for throttling
-                activity heartbeats in case per-activity heartbeat timeout is
-                unset. Otherwise, it's the per-activity heartbeat timeout * 0.8.
-            graceful_shutdown_timeout: Amount of time after shutdown is called
-                that activities are given to complete before their tasks are
-                cancelled.
-            shared_state_manager: Used for obtaining cross-process friendly
-                synchronization primitives. This is required for non-async
-                activities where the activity_executor is not a
-                :py:class:`concurrent.futures.ThreadPoolExecutor`. Reuse of
-                these across workers is encouraged.
-        """
-        # TODO(cretz): Support workflows
-        if not activities:
-            raise ValueError("At least one activity must be specified")
-        # If there are any non-async activities, an executor is required
+        # Validate and build activity dict
         self._activities: Dict[str, _ActivityDefinition] = {}
         for name, activity in activities.items():
             if not callable(activity):
@@ -180,193 +118,18 @@ class Worker:
                             f"Activity {name} must be picklable when using a process executor"
                         ) from err
             arg_types, ret_type = temporalio.converter._type_hints_from_func(
-                activity, eval_str=client._config["type_hint_eval_str"]
+                activity, eval_str=type_hint_eval_str
             )
             self._activities[name] = _ActivityDefinition(
                 name=name, fn=activity, arg_types=arg_types, ret_type=ret_type
             )
 
-        # Prepend applicable client interceptors to the given ones
-        interceptors_from_client = cast(
-            List[Interceptor],
-            [i for i in client.config()["interceptors"] if isinstance(i, Interceptor)],
-        )
-        interceptors = itertools.chain(interceptors_from_client, interceptors)
-
-        # Extract the bridge workflow service. We try the service on the client
-        # first, then we support a worker_workflow_service on the client's
-        # service to return underlying service we can use.
-        bridge_service: temporalio.workflow_service._BridgeWorkflowService
-        if isinstance(
-            client.service, temporalio.workflow_service._BridgeWorkflowService
-        ):
-            bridge_service = client.service
-        elif hasattr(client.service, "worker_workflow_service"):
-            bridge_service = client.service.worker_workflow_service
-            if not isinstance(
-                bridge_service, temporalio.workflow_service._BridgeWorkflowService
-            ):
-                raise TypeError(
-                    "Client service's worker_workflow_service cannot be used for a worker"
-                )
-        else:
-            raise TypeError(
-                "Client service cannot be used for a worker. "
-                + "Use the original client's service or set worker_workflow_service on the wrapped service with the original service."
-            )
-
-        self._bridge_worker = temporalio.bridge.worker.Worker(
-            bridge_service._bridge_client,
-            temporalio.bridge.worker.WorkerConfig(
-                namespace=client.namespace,
-                task_queue=task_queue,
-                max_cached_workflows=max_cached_workflows,
-                max_outstanding_workflow_tasks=max_concurrent_workflow_tasks,
-                max_outstanding_activities=max_concurrent_activities,
-                max_outstanding_local_activities=max_concurrent_local_activities,
-                max_concurrent_wft_polls=max_concurrent_wft_polls,
-                nonsticky_to_sticky_poll_ratio=nonsticky_to_sticky_poll_ratio,
-                max_concurrent_at_polls=max_concurrent_at_polls,
-                no_remote_activities=no_remote_activities,
-                sticky_queue_schedule_to_start_timeout_millis=int(
-                    1000 * sticky_queue_schedule_to_start_timeout.total_seconds()
-                ),
-                max_heartbeat_throttle_interval_millis=int(
-                    1000 * max_heartbeat_throttle_interval.total_seconds()
-                ),
-                default_heartbeat_throttle_interval_millis=int(
-                    1000 * default_heartbeat_throttle_interval.total_seconds()
-                ),
-            ),
-        )
-        # Store the config for tracking
-        self._config = WorkerConfig(
-            client=client,
-            task_queue=task_queue,
-            activities=activities,
-            activity_executor=activity_executor,
-            interceptors=interceptors,
-            max_cached_workflows=max_cached_workflows,
-            max_concurrent_workflow_tasks=max_concurrent_workflow_tasks,
-            max_concurrent_activities=max_concurrent_activities,
-            max_concurrent_local_activities=max_concurrent_local_activities,
-            max_concurrent_wft_polls=max_concurrent_wft_polls,
-            nonsticky_to_sticky_poll_ratio=nonsticky_to_sticky_poll_ratio,
-            max_concurrent_at_polls=max_concurrent_at_polls,
-            no_remote_activities=no_remote_activities,
-            sticky_queue_schedule_to_start_timeout=sticky_queue_schedule_to_start_timeout,
-            max_heartbeat_throttle_interval=max_heartbeat_throttle_interval,
-            default_heartbeat_throttle_interval=default_heartbeat_throttle_interval,
-            graceful_shutdown_timeout=graceful_shutdown_timeout,
-            shared_state_manager=shared_state_manager,
-        )
-        self._running_activities: Dict[bytes, _RunningActivity] = {}
-        self._task: Optional[asyncio.Task] = None
-        # Lazily created on first activity
-        self._worker_shutdown_event: Optional[
-            temporalio.activity._CompositeEvent
-        ] = None
-        self._seen_sync_activity = False
-
-    def config(self) -> WorkerConfig:
-        """Config, as a dictionary, used to create this worker.
-
-        Returns:
-            Configuration, shallow-copied.
-        """
-        config = self._config.copy()
-        config["activities"] = dict(config["activities"])
-        return config
-
-    @property
-    def task_queue(self) -> str:
-        """Task queue this worker is on."""
-        return self._config["task_queue"]
-
-    async def __aenter__(self) -> Worker:
-        """Start the worker and return self for use by ``async with``.
-
-        Returns:
-            Self.
-        """
-        self._start()
-        return self
-
-    async def __aexit__(self, *args) -> None:
-        """Same as :py:meth:`shutdown` for use by ``async with``."""
-        await self.shutdown()
-
     async def run(self) -> None:
-        """Run the worker and wait on it to be shutdown."""
-        await self._start()
-
-    def _start(self) -> asyncio.Task:
-        if self._task:
-            raise RuntimeError("Already started")
-        self._task = asyncio.create_task(
-            asyncio.wait([asyncio.create_task(self._run_activities())])
-        )
-        return self._task
-
-    async def shutdown(self) -> None:
-        """Shutdown the worker and wait until all activities have completed.
-
-        This will initiate a shutdown and optionally wait for a grace period
-        before sending cancels to all activities.
-        """
-        if not self._task:
-            raise RuntimeError("Never started")
-        graceful_timeout = self._config["graceful_shutdown_timeout"]
-        logger.info(
-            f"Beginning worker shutdown, will wait {graceful_timeout} before cancelling workflows/activities"
-        )
-        # Start shutdown of the bridge
-        bridge_shutdown_task = asyncio.create_task(self._bridge_worker.shutdown())
-        # Wait for the poller loop to stop
-        await self._task
-
-        # Collect all still running activity tasks
-        activity_tasks: List[asyncio.Task] = []
-        for activity in self._running_activities.values():
-            if activity.task:
-                activity_tasks.append(activity.task)
-        # Set event that we're shutting down (updates all activity tasks)
-        if self._worker_shutdown_event:
-            self._worker_shutdown_event.set()
-        if activity_tasks:
-            # Wait for any still running after graceful timeout
-            _, still_running = await asyncio.wait(
-                activity_tasks, timeout=graceful_timeout.total_seconds()
-            )
-            if still_running:
-                # Cancel all still running
-                if len(still_running) == 1:
-                    logger.info(f"Cancelling 1 activity that is still running")
-                else:
-                    logger.info(
-                        f"Cancelling {len(still_running)} activities that are still running"
-                    )
-                for task in still_running:
-                    # We have to find the running activity that's associated with
-                    # the task so that we can cancel through that. It's ok if the
-                    # activity is already gone.
-                    for activity in self._running_activities.values():
-                        if activity.info and activity.task is task:
-                            logger.info(
-                                f"Cancelling still-running activity: {activity.info._logger_details()}"
-                            )
-                            activity.cancel()
-                            break
-
-        # Wait for the bridge to report all activities are completed
-        await bridge_shutdown_task
-
-    async def _run_activities(self) -> None:
         # Continually poll for activity work
         while True:
             try:
                 # Poll for a task
-                task = await self._bridge_worker.poll_activity_task()
+                task = await self._bridge_worker().poll_activity_task()
 
                 if task.HasField("start"):
                     # Cancelled event and sync field will be updated inside
@@ -378,7 +141,7 @@ class Worker:
                     )
                     self._running_activities[task.task_token] = activity
                 elif task.HasField("cancel"):
-                    self._cancel_activity(task.task_token, task.cancel)
+                    self._cancel(task.task_token, task.cancel)
                 else:
                     raise RuntimeError(f"Unrecognized activity task: {task}")
             except temporalio.bridge.worker.PollShutdownError:
@@ -387,10 +150,44 @@ class Worker:
                 # Should never happen
                 logger.exception(f"Activity runner failed")
 
-    async def _run_workflows(self) -> None:
-        raise NotImplementedError
+    async def shutdown(self, after_graceful_timeout: timedelta) -> None:
+        # Set event that we're shutting down (updates all activity tasks)
+        if self._worker_shutdown_event:
+            self._worker_shutdown_event.set()
+        # Collect all still running activity tasks or exit if none
+        activity_tasks = [
+            activity.task
+            for activity in self._running_activities.values()
+            if activity.task
+        ]
+        if not activity_tasks:
+            return
+        # Wait for any still running after graceful timeout and exit if none
+        _, still_running = await asyncio.wait(
+            activity_tasks, timeout=after_graceful_timeout.total_seconds()
+        )
+        if not still_running:
+            return
+        # Cancel all still running
+        if len(still_running) == 1:
+            logger.info(f"Cancelling 1 activity that is still running")
+        else:
+            logger.info(
+                f"Cancelling {len(still_running)} activities that are still running"
+            )
+        for task in still_running:
+            # We have to find the running activity that's associated with
+            # the task so that we can cancel through that. It's ok if the
+            # activity is already gone.
+            for activity in self._running_activities.values():
+                if activity.info and activity.task is task:
+                    logger.info(
+                        f"Cancelling still-running activity: {activity.info._logger_details()}"
+                    )
+                    activity.cancel()
+                    break
 
-    def _cancel_activity(
+    def _cancel(
         self, task_token: bytes, cancel: temporalio.bridge.proto.activity_task.Cancel
     ) -> None:
         activity = self._running_activities.get(task_token)
@@ -400,7 +197,7 @@ class Worker:
         logger.debug("Cancelling activity %s, reason: %s", task_token, cancel.reason)
         activity.cancel(cancelled_by_request=True)
 
-    def _heartbeat_activity(self, task_token: bytes, *details: Any) -> None:
+    def _heartbeat(self, task_token: bytes, *details: Any) -> None:
         # We intentionally make heartbeating non-async, but since the data
         # converter is async, we have to schedule it. If the activity is done,
         # we do not schedule any more. Technically this should be impossible to
@@ -411,13 +208,13 @@ class Worker:
         activity = self._running_activities.get(task_token)
         if activity and not activity.done:
             # Put on queue and schedule a task. We will let the queue-full error
-            # be thrown here
+            # be thrown here.
             activity.pending_heartbeats.put_nowait(details)
             activity.last_heartbeat_task = asyncio.create_task(
-                self._heartbeat_activity_async(logger, activity, task_token)
+                self._heartbeat_async(logger, activity, task_token)
             )
 
-    async def _heartbeat_activity_async(
+    async def _heartbeat_async(
         self,
         logger: logging.LoggerAdapter,
         activity: _RunningActivity,
@@ -434,15 +231,13 @@ class Worker:
         try:
             heartbeat = temporalio.bridge.proto.ActivityHeartbeat(task_token=task_token)
             if details:
-                converted_details = await self._config["client"].data_converter.encode(
-                    details
-                )
+                converted_details = await self._data_converter.encode(details)
                 # Convert to core payloads
                 heartbeat.details.extend(
                     temporalio.bridge.worker.to_bridge_payloads(converted_details)
                 )
             logger.debug("Recording heartbeat with details %s", details)
-            self._bridge_worker.record_activity_heartbeat(heartbeat)
+            self._bridge_worker().record_activity_heartbeat(heartbeat)
         except Exception as err:
             # If the activity is done, nothing we can do but log
             if activity.done:
@@ -473,7 +268,7 @@ class Worker:
             # Find activity or fail
             activity_def = self._activities.get(start.activity_type)
             if not activity_def:
-                activity_names = ", ".join(sorted(self._config["activities"].keys()))
+                activity_names = ", ".join(sorted(self._activities.keys()))
                 raise temporalio.exceptions.ApplicationError(
                     f"Activity function {start.activity_type} is not registered on this worker, available activities: {activity_names}",
                     type="NotFoundError",
@@ -491,8 +286,7 @@ class Worker:
                 # If we're in a thread-pool executor we can use threading events
                 # otherwise we must use manager events
                 if isinstance(
-                    self._config["activity_executor"],
-                    concurrent.futures.ThreadPoolExecutor,
+                    self._activity_executor, concurrent.futures.ThreadPoolExecutor
                 ):
                     running_activity.cancelled_event = temporalio.activity._CompositeEvent(
                         thread_event=threading.Event(),
@@ -500,7 +294,7 @@ class Worker:
                         async_event=None,
                     )
                 else:
-                    manager = self._config["shared_state_manager"]
+                    manager = self._shared_state_manager
                     # Pre-checked on worker init
                     assert manager
                     running_activity.cancelled_event = temporalio.activity._CompositeEvent(
@@ -529,12 +323,11 @@ class Worker:
                 activity_def.arg_types
             ) != len(start.input):
                 arg_types = None
-            converter = self._config["client"].data_converter
             try:
                 args = (
                     []
                     if not start.input
-                    else await converter.decode(
+                    else await self._data_converter.decode(
                         temporalio.bridge.worker.from_bridge_payloads(start.input),
                         type_hints=arg_types,
                     )
@@ -550,7 +343,7 @@ class Worker:
                 heartbeat_details = (
                     []
                     if not start.heartbeat_details
-                    else await converter.decode(
+                    else await self._data_converter.decode(
                         temporalio.bridge.worker.from_bridge_payloads(
                             start.heartbeat_details
                         )
@@ -591,7 +384,7 @@ class Worker:
                 if start.HasField("start_to_close_timeout")
                 else None,
                 started_time=_proto_to_datetime(start.started_time),
-                task_queue=self._config["task_queue"],
+                task_queue=self._task_queue,
                 task_token=task_token,
                 workflow_id=start.workflow_execution.workflow_id,
                 workflow_namespace=start.workflow_namespace,
@@ -602,12 +395,8 @@ class Worker:
             input = ExecuteActivityInput(
                 fn=activity_def.fn,
                 args=args,
-                executor=None
-                if not running_activity.sync
-                else self._config["activity_executor"],
+                executor=None if not running_activity.sync else self._activity_executor,
                 _cancelled_event=running_activity.cancelled_event,
-                _worker_shutdown_event=self._worker_shutdown_event,
-                _worker=self,
             )
 
             # Set the context early so the logging adapter works and
@@ -625,8 +414,8 @@ class Worker:
             # Build the interceptors chaining in reverse. We build a context right
             # now even though the info() can't be intercepted and heartbeat() will
             # fail. The interceptors may want to use the info() during init.
-            impl: ActivityInboundInterceptor = _ActivityInboundImpl()
-            for interceptor in reversed(list(self._config["interceptors"])):
+            impl: ActivityInboundInterceptor = _ActivityInboundImpl(self)
+            for interceptor in reversed(list(self._interceptors)):
                 impl = interceptor.intercept_activity(impl)
             # Init
             impl.init(_ActivityOutboundImpl(self, running_activity.info))
@@ -638,9 +427,7 @@ class Worker:
             if result is None:
                 completion.result.completed.SetInParent()
             else:
-                result_payloads = await self._config["client"].data_converter.encode(
-                    [result]
-                )
+                result_payloads = await self._data_converter.encode([result])
                 completion.result.completed.result.CopyFrom(
                     temporalio.bridge.worker.to_bridge_payload(result_payloads[0])
                 )
@@ -659,7 +446,7 @@ class Worker:
                     )
                     await temporalio.exceptions.apply_exception_to_failure(
                         err,
-                        self._config["client"].data_converter,
+                        self._data_converter,
                         completion.result.failed.failure,
                     )
                 elif (
@@ -670,7 +457,7 @@ class Worker:
                     await temporalio.exceptions.apply_error_to_failure(
                         # TODO(cretz): Should use some other message?
                         temporalio.exceptions.CancelledError("Cancelled"),
-                        self._config["client"].data_converter,
+                        self._data_converter,
                         completion.result.cancelled.failure,
                     )
                 else:
@@ -679,7 +466,7 @@ class Worker:
                     )
                     await temporalio.exceptions.apply_exception_to_failure(
                         err,
-                        self._config["client"].data_converter,
+                        self._data_converter,
                         completion.result.failed.failure,
                     )
             except Exception as inner_err:
@@ -708,32 +495,9 @@ class Worker:
             # Send task completion to core
             del self._running_activities[task_token]
             logger.debug("Completing activity with completion: %s", completion)
-            await self._bridge_worker.complete_activity_task(completion)
+            await self._bridge_worker().complete_activity_task(completion)
         except Exception:
             temporalio.activity.logger.exception("Failed completing activity task")
-
-
-class WorkerConfig(typing_extensions.TypedDict, total=False):
-    """TypedDict of config originally passed to :py:class:`Worker`."""
-
-    client: temporalio.client.Client
-    task_queue: str
-    activities: Mapping[str, Callable]
-    activity_executor: Optional[concurrent.futures.Executor]
-    interceptors: Iterable[Interceptor]
-    max_cached_workflows: int
-    max_concurrent_workflow_tasks: int
-    max_concurrent_activities: int
-    max_concurrent_local_activities: int
-    max_concurrent_wft_polls: int
-    nonsticky_to_sticky_poll_ratio: float
-    max_concurrent_at_polls: int
-    no_remote_activities: bool
-    sticky_queue_schedule_to_start_timeout: timedelta
-    max_heartbeat_throttle_interval: timedelta
-    default_heartbeat_throttle_interval: timedelta
-    graceful_shutdown_timeout: timedelta
-    shared_state_manager: Optional[SharedStateManager]
 
 
 @dataclass
@@ -773,97 +537,10 @@ class _RunningActivity:
             self.task.cancel()
 
 
-@dataclass
-class ExecuteActivityInput:
-    """Input for :py:meth:`ActivityInboundInterceptor.execute_activity`."""
-
-    fn: Callable[..., Any]
-    args: Iterable[Any]
-    executor: Optional[concurrent.futures.Executor]
-    _cancelled_event: temporalio.activity._CompositeEvent
-    _worker_shutdown_event: temporalio.activity._CompositeEvent
-    _worker: Worker
-
-
-class Interceptor:
-    """Interceptor for workers.
-
-    This should be extended by any worker interceptors.
-    """
-
-    def intercept_activity(
-        self, next: ActivityInboundInterceptor
-    ) -> ActivityInboundInterceptor:
-        """Method called for intercepting an activity.
-
-        Args:
-            next: The underlying inbound interceptor this interceptor should
-                delegate to.
-
-        Returns:
-            The new interceptor that will be used to for the activity.
-        """
-        return next
-
-
-class ActivityInboundInterceptor:
-    """Inbound interceptor to wrap outbound creation and activity execution.
-
-    This should be extended by any activity inbound interceptors.
-    """
-
-    def __init__(self, next: ActivityInboundInterceptor) -> None:
-        """Create the inbound interceptor.
-
-        Args:
-            next: The next interceptor in the chain. The default implementation
-                of all calls is to delegate to the next interceptor.
-        """
-        self.next = next
-
-    def init(self, outbound: ActivityOutboundInterceptor) -> None:
-        """Initialize with an outbound interceptor.
-
-        To add a custom outbound interceptor, wrap the given interceptor before
-        sending to the next ``init`` call.
-        """
-        self.next.init(outbound)
-
-    async def execute_activity(self, input: ExecuteActivityInput) -> Any:
-        """Called to invoke the activity."""
-        return await self.next.execute_activity(input)
-
-
-class ActivityOutboundInterceptor:
-    """Outbound interceptor to wrap calls made from within activities.
-
-    This should be extended by any activity outbound interceptors.
-    """
-
-    def __init__(self, next: ActivityOutboundInterceptor) -> None:
-        """Create the outbound interceptor.
-
-        Args:
-            next: The next interceptor in the chain. The default implementation
-                of all calls is to delegate to the next interceptor.
-        """
-        self.next = next
-
-    def info(self) -> temporalio.activity.Info:
-        """Called for every :py:func:`temporalio.activity.info` call."""
-        return self.next.info()
-
-    def heartbeat(self, *details: Any) -> None:
-        """Called for every :py:func:`temporalio.activity.heartbeat` call."""
-        self.next.heartbeat(*details)
-
-    # TODO(cretz): Do we want outbound interceptors for other items?
-
-
 class _ActivityInboundImpl(ActivityInboundInterceptor):
-    def __init__(self) -> None:
+    def __init__(self, worker: _Worker) -> None:
         # We are intentionally not calling the base class's __init__ here
-        pass
+        self._worker = worker
 
     def init(self, outbound: ActivityOutboundInterceptor) -> None:
         # Set the context callables. We are setting values instead of replacing
@@ -909,13 +586,16 @@ class _ActivityInboundImpl(ActivityInboundInterceptor):
             shared_manager: Optional[SharedStateManager] = None
             if not isinstance(input.executor, concurrent.futures.ThreadPoolExecutor):
                 # Should always be present in worker, pre-checked on init
-                shared_manager = input._worker._config["shared_state_manager"]
+                shared_manager = self._worker._shared_state_manager
                 assert shared_manager
                 heartbeat = await shared_manager.register_heartbeater(
                     info.task_token, ctx.heartbeat
                 )
 
             try:
+                # Shutdown event always present here
+                shutdown_event = self._worker._worker_shutdown_event
+                assert shutdown_event
                 return await loop.run_in_executor(
                     input.executor,
                     _execute_sync_activity,
@@ -923,7 +603,7 @@ class _ActivityInboundImpl(ActivityInboundInterceptor):
                     heartbeat,
                     # Only thread event, this may cross a process boundary
                     input._cancelled_event.thread_event,
-                    input._worker_shutdown_event.thread_event,
+                    shutdown_event.thread_event,
                     input.fn,
                     *input.args,
                 )
@@ -933,6 +613,20 @@ class _ActivityInboundImpl(ActivityInboundInterceptor):
 
         # Otherwise for async activity, just run
         return await input.fn(*input.args)
+
+
+class _ActivityOutboundImpl(ActivityOutboundInterceptor):
+    def __init__(self, worker: _Worker, info: temporalio.activity.Info) -> None:
+        # We are intentionally not calling the base class's __init__ here
+        self._worker = worker
+        self._info = info
+
+    def info(self) -> temporalio.activity.Info:
+        return self._info
+
+    def heartbeat(self, *details: Any) -> None:
+        info = temporalio.activity.info()
+        self._worker._heartbeat(info.task_token, *details)
 
 
 # This has to be defined at the top-level to be picklable for process executors
@@ -966,27 +660,6 @@ def _execute_sync_activity(
         )
     )
     return fn(*args)
-
-
-class _ActivityOutboundImpl(ActivityOutboundInterceptor):
-    def __init__(self, worker: Worker, info: temporalio.activity.Info) -> None:
-        # We are intentionally not calling the base class's __init__ here
-        self._worker = worker
-        self._info = info
-
-    def info(self) -> temporalio.activity.Info:
-        return self._info
-
-    def heartbeat(self, *details: Any) -> None:
-        info = temporalio.activity.info()
-        self._worker._heartbeat_activity(info.task_token, *details)
-
-
-def _proto_to_datetime(
-    ts: google.protobuf.timestamp_pb2.Timestamp,
-) -> datetime:
-    # Protobuf doesn't set the timezone but we want to
-    return ts.ToDatetime().replace(tzinfo=timezone.utc)
 
 
 class SharedStateManager(ABC):
@@ -1146,3 +819,10 @@ class _MultiprocessingSharedHeartbeatSender(SharedHeartbeatSender):
         # timeout after 30 seconds (should be plenty if the queue is being
         # properly processed)
         self._heartbeat_queue.put((task_token, details), True, 30)
+
+
+def _proto_to_datetime(
+    ts: google.protobuf.timestamp_pb2.Timestamp,
+) -> datetime:
+    # Protobuf doesn't set the timezone but we want to
+    return ts.ToDatetime().replace(tzinfo=timezone.utc)
